Move a few more components
