+++ src/org/labkey/test/components/ui/entities/EntityBulkUpdateDialog.java
@@ -6,7 +6,7 @@ import org.labkey.test.components.bootstrap.ModalDialog;
 import org.labkey.test.components.react.FilteringReactSelect;
 import org.labkey.test.components.html.Checkbox;
 import org.labkey.test.components.html.Input;
-import org.labkey.test.components.html.ToggleButton;
+import org.labkey.test.components.react.ToggleButton;
 import org.labkey.test.components.react.ReactDatePicker;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
