+++ src/org/labkey/test/components/ui/grids/GridBar.java
@@ -9,7 +9,7 @@ import org.labkey.test.Locator;
 import org.labkey.test.components.Component;
 import org.labkey.test.components.WebDriverComponent;
 import org.labkey.test.components.react.MultiMenu;
-import org.labkey.test.components.react.Pager;
+import org.labkey.test.components.ui.Pager;
 import org.openqa.selenium.ElementNotVisibleException;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.StaleElementReferenceException;
