+++ src/org/labkey/test/components/ui/grids/ResponsiveGrid.java
@@ -11,7 +11,7 @@ import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.components.Component;
 import org.labkey.test.components.WebDriverComponent;
 import org.labkey.test.components.react.ReactCheckBox;
-import org.labkey.test.components.react.UpdatingComponent;
+import org.labkey.test.components.UpdatingComponent;
 import org.openqa.selenium.NotFoundException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
